# [esp32] Enable LWIP core locking on ESP-IDF to reduce socket operation overhead