# What does this implement/fix?

This PR enables LWIP TCP/IP core locking by default on ESP-IDF to significantly improve socket operation performance. This brings ESP-IDF socket performance to parity with Arduino framework, which already has this enabled.

### Background
When switching from Arduino to ESP-IDF framework, I noticed a significant performance drop in network operations. Investigation revealed that ESP-IDF's default LWIP configuration uses mailbox message passing for all socket operations, while Arduino uses direct calls with core locking.

### Performance Impact
Benchmarks (below) show that socket operations are 24-200% slower without core locking:
- `select()` on 4 sockets: ~190μs (Arduino) vs ~235μs (ESP-IDF default) - 24% overhead
- Under system load, the overhead can reach 100-200% as operations queue in the tcpip_thread mailbox

**Detailed benchmark results with core locking enabled on ESP-IDF:**

Before (ESP-IDF default - mailbox/context switching):
- `select()` on 4 sockets: ~235μs
- `recv()` on 4 sockets: ~130μs

After (with core locking enabled):
- `select()` on 4 sockets: ~169μs (28% faster)
- `recv()` on 4 sockets: ~123μs

The improvement matches and even slightly exceeds Arduino framework performance, while eliminating the context switching overhead.

This overhead accumulates significantly for components that make frequent socket operations like the API server, web server, and MQTT.

### Solution
This PR adds two new ESP-IDF advanced configuration options (defaulting to enabled):
- `enable_lwip_tcpip_core_locking`: Enables direct socket calls with locking instead of mailbox message passing
- `enable_lwip_check_thread_safety`: Enables thread safety checking to catch potential issues

Since Arduino framework already has core locking enabled by default and ESPHome must support it anyway, enabling it for ESP-IDF improves performance consistency across frameworks.

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):** N/A

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):** esphome/esphome-docs#5169

## Test Environment

- [ ] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx
- [ ] nRF52840

## Example entry for `config.yaml`:

```yaml
# Example config.yaml
esp32:
  board: esp32dev
  framework:
    type: esp-idf
    advanced:
      # These are now enabled by default for better performance
      # Disable only if you need the multi-threaded scalability of mailbox approach
      enable_lwip_tcpip_core_locking: true  # Default: true
      enable_lwip_check_thread_safety: true  # Default: true
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).