@bdraco 2025-07-24
I also need to test with the wrong password for wifi